Presentation Slides
These are the presentation slides that I used when presenting my findings to non-technical stakeholders at the project event. These slides showcase the visualisations I made using Matplotlib and Seaborn and are not text heavy. This approach worked well as it allowed me to use the graphs to tell a story, as opposed to explaining raw numbers, which resulted in clear insights for stakeholders that could go on to affect station maintenance patterns and resource allocation in the future.